+++ src/component/aboutme/skillTags.js
@@ -1,39 +1,10 @@
 import React from "react";
-import { makeStyles } from "@material-ui/core/styles";
-import Chip from "@material-ui/core/Chip";
+import { useStyles } from "./style";
+import { Chip } from "@material-ui/core";
 
-const useStyles = makeStyles(theme => ({
-  tag: {
-    display: "flex",
-    justifyContent: "left",
-    flexWrap: "wrap",
-    marginTop: "10px",
-    "& > *": {
-      margin: theme.spacing(0.6)
-    }
-  }
-}));
 
-const SkillTagsAbm = () => {
+const SkillTagsAbm = ({ skills }) => {
   const classes = useStyles();
-
-  const skills = [
-    "HTML/CSS",
-    "PHP",
-    "JavaScript",
-    "Java",
-    "ReactJS",
-    "Python",
-    "C",
-    "Assembly",
-    "CCNA certificated",
-    "Software Defined Network",
-    "AWS",
-    "Ardunio",
-    "Zigbee",
-    "RFID",
-    "Wireless AP/Router"
-  ];
 
   return (
     <div className={classes.tag}>
