+++ src/component/aboutme/skillsData.js
@@ -1,0 +1,19 @@
+const skills = [
+    "HTML/CSS",
+    "PHP",
+    "JavaScript",
+    "Java",
+    "ReactJS",
+    "Python",
+    "C",
+    "Assembly",
+    "CCNA certificated",
+    "Software Defined Network",
+    "AWS",
+    "Ardunio",
+    "Zigbee",
+    "RFID",
+    "Wireless AP/Router"
+];
+
+export default skills;
