+++ src/component/nav/Nav.js
@@ -1,32 +1,8 @@
 import React from "react";
-import { makeStyles } from "@material-ui/core/styles";
-import AppBar from "@material-ui/core/AppBar";
-import Toolbar from "@material-ui/core/Toolbar";
-import Typography from "@material-ui/core/Typography";
-import Button from "@material-ui/core/Button";
-import pink from "@material-ui/core/colors/pink";
-import Link from "@material-ui/core/Link";
-import Container from "@material-ui/core/Container";
+import { useStyles } from "./style";
+import { AppBar, Toolbar, Typography, Button, Link, Container } from "@material-ui/core";
 
 
-const useStyles = makeStyles(theme => ({
-  nav_block: {
-    flexGrow: 1
-  },
-  nav_bar: {
-    background: "#fafafa",
-    boxShadow: "none"
-  },
-  nav_lastL: {
-    color: pink[400],
-    fontweight: "bold"
-  },
-  nav_logo: {
-    flexGrow: 1,
-    color: "black",
-    fontWeight: "bold",
-  }
-}));
 
 
 const Nav = ({ abtRef, pjRef }) => {
